% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_and_tidy_date}
\alias{check_and_tidy_date}
\title{Validate general dates}
\usage{
check_and_tidy_date(date, date_name)
}
\arguments{
\item{date}{A date object in string format
such as '2018-01-01'.}

\item{date_name}{The name of the date variable.}
}
\value{
A verified date object in string format
}
\description{
A function to check the date variable whether 
is the right date. This is an internal function.
}
