% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_and_tidy_date_crossing}
\alias{check_and_tidy_date_crossing}
\title{Validate dates}
\usage{
check_and_tidy_date_crossing(date_crossing)
}
\arguments{
\item{date_crossing}{The border crossing date.}
}
\description{
Validate dates
}
