% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_ini_res_data_columns}
\alias{check_ini_res_data_columns}
\title{Validate the data columns}
\usage{
check_ini_res_data_columns(data)
}
\arguments{
\item{data}{The journey data that should contain columns of (
'personId',
'res_status_initial',
'date_finalised').}
}
\description{
Validate the data columns
}
