% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_is_logic}
\alias{check_is_logic}
\title{Validate a logical value}
\usage{
check_is_logic(check_value)
}
\arguments{
\item{check_value}{Boolean value to present In/Out the country.}
}
\value{
A NULL value if there is no issue raised.
}
\description{
A function to check the variable whether 
is the right logic type. This is an internal function.
}
