% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_positive_number}
\alias{check_positive_number}
\title{Validate a positive numeric value}
\usage{
check_positive_number(number, name)
}
\arguments{
\item{number}{The checking value.}

\item{name}{The name of the variable.}
}
\value{
A NULL value if there is no issue raised.
}
\description{
A function to check the variable whether is positive number. 
This is an internal function.
}
