% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_object_size}
\alias{get_object_size}
\title{Get Object Size}
\usage{
get_object_size(object)
}
\arguments{
\item{object}{The target object.}
}
\value{
A named list object that contains information on the 
 size of an object and the size unit.
}
\description{
A function to get the size of an object
}
\examples{
res <- get_object_size(TRUE)
res$size
res$unit

}
