% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_data.R
\name{internal_process}
\alias{internal_process}
\title{Internal function}
\usage{
internal_process(subgroup, window_size, threshold_year)
}
\arguments{
\item{subgroup}{A subgroup of the pre-processed data groups,
generated by \code{migrbc::pre_process}.}

\item{window_size}{The maximum length of the scanning period.
Can be an integer giving the number of days, the result
of a call to function \code{\link[base]{difftime}}, or an object of
class \code{\link[lubridate:Duration-class]{Duration}}.}

\item{threshold_year}{The length of the yearly test period.
Can be an integer giving the number of days, the result
of a call to function \code{\link[base]{difftime}}, or an object of
class \code{\link[lubridate:Duration-class]{Duration}}.}
}
\value{
A data frame object of classified / labelled journeys.
}
\description{
Internal function
}
