% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_resolve}
\alias{rcpp_resolve}
\title{Processing RBC for a person.}
\usage{
rcpp_resolve(person_data, int_res_status, initial_date_finalised, tw, tm)
}
\arguments{
\item{person_data}{A list object of a person's journeys.}

\item{int_res_status}{The initial residence status of the target person}

\item{initial_date_finalised}{The final resolved date of the initial residence
status.}

\item{tw}{Windows Size, by default, it is 487 days.}

\item{tm}{Threshold of Year, by default, it is 365 days.}
}
\value{
A list of classified / labelled journeys.
}
\description{
This function is used to resolve one person's journeys, i.e., classifying 
a person and marking it whether or not to be a long term migrant based on 
the person's journeys. This function is used internally 
inside the package and shouldn't be exposed to the outside caller.
}
