% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_rbc_process_core}
\alias{run_rbc_process_core}
\title{Processing RBC for a list of person.}
\usage{
run_rbc_process_core(cross_data, ini_status_data, tw, ty)
}
\arguments{
\item{cross_data}{The personal crossing data for RBC process}

\item{ini_status_data}{the initial residence status data}

\item{tw}{Windows Size, by default, it is 487 days.}

\item{ty}{Threshold of Year, by default, it is 365 days.}
}
\value{
A DataFrame object
}
\description{
This function is used to resolve a list of person's journeys,
i.e., classifying a list of people and marking it whether or not 
to be a long term migrant based on the person's journeys. 
This function is used internally inside the package and shouldn't 
be exposed to the outside caller.
}
