% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_rbc_process_with_error}
\alias{run_rbc_process_with_error}
\title{Processing RBC for a list of person.}
\usage{
run_rbc_process_with_error(cross_data, ini_status_data, error_message, tw)
}
\arguments{
\item{cross_data}{The personal crossing data for RBC process}

\item{ini_status_data}{the initial residence status data}

\item{error_message}{The error message.}

\item{tw}{Windows Size, by default, it is 487 days.}
}
\value{
A data frame object of classified / labelled journeys
}
\description{
This function is used to resolve a list of person's journeys with error,
This function is used internally inside the package and shouldn't 
be exposed to the outside caller.
}
