% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mig_hist.R
\name{segment_coord_horiz}
\alias{segment_coord_horiz}
\title{Internal function}
\usage{
segment_coord_horiz(date_crossing, is_arrival, date_first, date_last)
}
\arguments{
\item{date_crossing}{date of border crossing.}

\item{is_arrival}{A Boolean value.}

\item{date_first}{The first date occurred.}

\item{date_last}{The last date occurred.}
}
\value{
A list object that contains values for coordinates
 of horizon.
}
\description{
Internal function
}
