% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mig_hist.R
\name{segment_coord_vert}
\alias{segment_coord_vert}
\title{Internal function}
\usage{
segment_coord_vert(date_crossing, is_arrival)
}
\arguments{
\item{date_crossing}{date of border crossing.}

\item{is_arrival}{A Boolean value.}
}
\description{
Internal function
}
