% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_and_tidy_date_crossing}
\alias{check_and_tidy_date_crossing}
\title{Validate dates on border crossing.}
\usage{
check_and_tidy_date_crossing(date_crossing)
}
\arguments{
\item{date_crossing}{The border crossing date.}
}
\value{
The border crossing date that has been verified and 
tidied up.
}
\description{
A function to check the date variable whether 
is the right date. This is an internal function.
}
