% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_and_tidy_date_first_last}
\alias{check_and_tidy_date_first_last}
\title{Validate dates}
\usage{
check_and_tidy_date_first_last(date, date_crossing, name)
}
\arguments{
\item{date}{The last date to compare with.}

\item{date_crossing}{The border crossing date.}

\item{name}{A name of the checking variable.}
}
\description{
Validate dates
}
