% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_data_columns}
\alias{check_data_columns}
\title{Validate the data columns}
\usage{
check_data_columns(data)
}
\arguments{
\item{data}{The journey data that should contain columns in (
'journeyId', 
'personId', 
'date_crossing', 
'is_arrival',
'journey_sequence',
'journeyId_prev).}
}
\description{
Validate the data columns
}
