% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_integer}
\alias{check_integer}
\title{Validate an integer value}
\usage{
check_integer(name = NULL, value = NULL)
}
\arguments{
\item{name}{The name of the variable.}

\item{value}{The validating variable.}
}
\value{
A NULL value if there is no issue raised.
}
\description{
A function to check the variable whether 
is the right integer type. This is an internal function.
}
