% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_is_logic}
\alias{check_is_logic}
\title{Validate a logical value}
\usage{
check_is_logic(check_value)
}
\arguments{
\item{check_value}{Boolean value to present In/Out the country.}
}
\description{
Check a value whether or not it is a Boolean value.
}
