% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_positive_number}
\alias{check_positive_number}
\title{Validate a numeric value}
\usage{
check_positive_number(number, name)
}
\arguments{
\item{number}{The checking value.}

\item{name}{The name of the variable.}
}
\description{
Check a value whether or not it is a number.
}
