% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_object_size}
\alias{get_object_size}
\title{A function to check the size of an object}
\usage{
get_object_size(object)
}
\arguments{
\item{object}{The target object.}
}
\value{
A named list object.
}
\description{
A function to check the size of an object
}
\examples{
res <- get_object_size(TRUE)
res$size
res$unit

}
