% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{setup_random_test_data}
\alias{setup_random_test_data}
\title{A function to generate test data for RBC}
\usage{
setup_random_test_data(num_people = 1000, initial_date = "2001-01-01",
  numJourneys = 5, min = 0, max = 10)
}
\arguments{
\item{num_people}{The number of person instances.}

\item{initial_date}{The start crossing date.}

\item{numJourneys}{The number of journeys for each person.}

\item{min}{The minimum duration between journeys.}

\item{max}{The maximum duration between journeys.}
}
\value{
A data frame object
}
\description{
A function to generate test data for RBC
}
\examples{

res <- setup_random_test_data(1000, 
                              initial_date = '2001-01-01',
                              numJourneys = 5,
                              min = 0,
                              max = 10)
res

}
